requirements.txt
aioredis==2.0.1
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.85.0
h11==0.12.0
hiredis==2.0.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
kombu==5.2.4
lxml==4.9.1
packaging==21.3
prompt-toolkit==3.0.31
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
vine==5.0.0
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
xmltodict==0.13.0
sqlalchemy
sqlmodel
alembic
